Take care of PR feedbacks
